ubusd: make txq_len field signed
When enqueueing a partial message, the length is initially set to a
negative value. Fixes part of the enqueue queue length check.

+++ ubusd.h
@@ -62,7 +62,7 @@ struct ubus_client {
 	struct list_head cmd_queue;
 	struct list_head tx_queue;
 	unsigned int txq_ofs;
-	unsigned int txq_len;
+	ssize_t txq_len;
 
 	struct ubus_msg_buf *pending_msg;
 	struct ubus_msg_buf *retmsg;
